fix: unfocus URL field, request drawer close
