+++ lib/screens/home_page/collection_pane.dart
@@ -158,6 +158,7 @@ class RequestItem extends ConsumerWidget {
   Widget build(BuildContext context, WidgetRef ref) {
     final selectedId = ref.watch(selectedIdStateProvider);
     final editRequestId = ref.watch(selectedIdEditStateProvider);
+    final mobileDrawerKey = ref.watch(mobileDrawerKeyProvider);
 
     return SidebarRequestCard(
       id: id,
@@ -167,6 +168,7 @@ class RequestItem extends ConsumerWidget {
       selectedId: selectedId,
       editRequestId: editRequestId,
       onTap: () {
+        mobileDrawerKey.currentState?.close();
         ref.read(selectedIdStateProvider.notifier).state = id;
       },
       // onDoubleTap: () {
