+++ lib/screens/home_page/editor_pane/editor_pane.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:apidash/providers/providers.dart';
-import 'package:apidash/consts.dart';
 import 'editor_default.dart';
 import 'editor_request.dart';
 
@@ -16,10 +15,7 @@ class RequestEditorPane extends ConsumerWidget {
     if (selectedId == null) {
       return const RequestEditorDefault();
     } else {
-      return Padding(
-        padding: kIsMacOS || kIsWindows ? kPt24o8 : kP8,
-        child: const RequestEditor(),
-      );
+      return const RequestEditor();
     }
   }
 }
